+++ src/index.html
@@ -25,7 +25,7 @@
             </form>
             
     </section>
-    <script src="index.js" ></script>
+    <script type="module" src="index.js" ></script>
     
 </body>
 </html>
